+++ turbo/apps/cli/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [9.49.2](https://github.com/vm0-ai/vm0/compare/cli-v9.49.1...cli-v9.49.2) (2026-03-06)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @vm0/core bumped to 8.42.0
+
 ## [9.49.1](https://github.com/vm0-ai/vm0/compare/cli-v9.49.0...cli-v9.49.1) (2026-03-06)